tests/wpt/web-platform-tests/websockets/Create-nonAscii-protocol-string.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - non ascii protocol string</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        if(window.WebSocket) {
            test(function () {
                var nonAsciiProtocol = "\u0080echo";
                var wsocket;
                assert_throws("SYNTAX_ERR", function () { wsocket = CreateWebSocketNonAsciiProtocol(nonAsciiProtocol) });
            }, "W3C WebSocket API - Create WebSocket - Pass a valid URL and a protocol string with non-ascii values - SYNTAX_ERR is thrown")
        }
    </script>
</body>
</html>
